+++ docs/results.js
@@ -28,4 +28,12 @@
  * @memberof Realm
  */
 class Results extends Collection {
+    /**
+     * Bulk update objects in the collection.
+     * @param {string} property - The name of the property.
+     * @param {string} value - The updated property value.
+     * @throws {Error} If no property with the name exists.
+     * @since 1.12.1
+     */
+    update(property, value) {}
 }